Converted the project to use Webpack WIP
