+++ .vscode/settings.json
@@ -2,6 +2,7 @@
     "cSpell.words": [
         "Crosspost",
         "Kbps",
+        "mdfr",
         "shreddit"
     ]
 }
