+++ public/manifest.json
@@ -1,7 +1,7 @@
 {
   "manifest_version": 2,
   "name": "Media Downloader for Reddit",
-  "version": "1.4.2",
+  "version": "1.5.0",
   "description": "Makes it easy to download media content from Reddit.",
   "homepage_url": "https://github.com/ArualDev/media-downloader-for-reddit",
   "browser_specific_settings": {
@@ -34,7 +34,7 @@
         "*://*.reddit.com/*"
       ],
       "js": [
-        "init.js"
+        "content.js"
       ],
       "css": [
         "styles.css"
