+++ public/options.html
@@ -10,9 +10,9 @@
 
 <body>
     <form id="options-form">
-        <label for="enable-logging-cb">
-            Enable logging
-            <input type="checkbox" name="enable-logging" id="enable-logging-cb">
+        <label for="download-path-tx">
+            Download path
+            <input type="text" pattern="(\/[\w\s-]*)+" name="download-path" id="download-path-tx">
         </label>
         <label for="use-custom-server-cb">
             Use a custom server
@@ -22,6 +22,10 @@
             Custom server address
             <input type="text" pattern="https?:\/\/[a-z0-9]+(:\d{1,5})?" name="custom-server-address"
                 id="custom-server-address-tx">
+        </label>
+        <label for="enable-logging-cb">
+            Enable logging
+            <input type="checkbox" name="enable-logging" id="enable-logging-cb">
         </label>
 
         <div id="button-container">
