+++ src/background.js
@@ -1,19 +1,14 @@
-function downloadFromUrl(url, filename, saveAs = false) {
-	browser.downloads.download({
-		url: url,
-		filename: filename,
-		saveAs: saveAs
-	});
-}
-
 const messageActions = {
 	download: msg => {
-		downloadFromUrl(msg.url, msg.filename, msg.saveAs)
+		browser.downloads.download({
+			url: msg.url,
+			filename: msg.filename,
+			saveAs: msg.saveAs
+		});
 	},
 	getVersion: (message, sender, sendResponse) => {
 		const manifest = browser.runtime.getManifest();
-		const extensionVersion = manifest.version;
-		sendResponse(extensionVersion);
+		sendResponse(manifest.version);
 	}
 }
