+++ src/classes/PostData.js
@@ -1,0 +1,7 @@
+export default class PostData {
+    constructor(url, title, downloads) {
+        this.url = url;
+        this.title = title;
+        this.downloads = downloads;
+    }
+}
