+++ src/handlers/handleFeedNewUI.js
@@ -1,0 +1,4 @@
+// TODO: Implement
+export default function handleFeedNewUI() {
+    throw new Error('Not Implemented');
+}
