+++ src/helpers/constants.js
@@ -1,0 +1,8 @@
+export const REFRESH_INTERVAL_MS = 500;
+export const REDDIT_IMAGE_EXTENSIONS = [".jpg", ".jpeg", ".png", ".gif", ".webp"];
+export const DEFAULT_OPTIONS = {
+    downloadPath: '/',
+    enableLogging: false,
+    useCustomServer: false,
+    customServerAddress: 'http://localhost:21370'
+}
