+++ src/options.js
@@ -1,48 +1,57 @@
+import { DEFAULT_OPTIONS } from "./helpers/constants";
+
+const optionsFrom = document.querySelector('#options-form');
+
 const enableLoggingCB = document.querySelector('#enable-logging-cb');
 const useCustomServerCB = document.querySelector('#use-custom-server-cb');
 const customServerAddressTx = document.querySelector('#custom-server-address-tx');
-const optionsFrom = document.querySelector('#options-form');
+const downloadPathTx = document.querySelector('#download-path-tx');
 
 const resetDefaultBtn = document.querySelector('#reset-default-btn');
 
-const defaultOptions = {
-    enableLogging: false,
-    useCustomServer: false,
-    customServerAddress: 'http://localhost:21370'
-}
-
 function setDefaultsIfUndefined(options) {
     let modified = false;
-    for (const k in defaultOptions) {
+    for (const k in DEFAULT_OPTIONS) {
         if (!options.hasOwnProperty(k)) {
-            options[k] = defaultOptions[k]
+            options[k] = DEFAULT_OPTIONS[k]
             modified = true;
         }
     }
     return modified;
 }
 
-browser.storage.sync.get('options').then(async res => {
-    const options = res.options ?? {};
-    const modified = setDefaultsIfUndefined(options);
-    if (modified) {
-        await browser.storage.sync.set({ options: options })
-    }
-    loadOptionsToDom(options);
-    updateCustomServerAddressVisibility();
-});
+async function loadOptions() {
+    return new Promise((resolve, reject) => {
+        browser.storage.sync.get('options').then(async res => {
+            const options = res.options ?? {};
+            const modified = setDefaultsIfUndefined(options);
+            if (modified) {
+                await browser.storage.sync.set({ options: options })
+            }
+            resolve(options)
+        });
+    });
+}
+
+loadOptions()
+    .then(options => {
+        loadOptionsToDom(options);
+        updateCustomServerAddressVisibility();
+    });
 
 async function saveOptionsFromDom() {
     const options = {
+        downloadPath: downloadPathTx.value,
         enableLogging: enableLoggingCB.checked,
         useCustomServer: useCustomServerCB.checked,
         customServerAddress: customServerAddressTx.value
     };
-    await browser.storage.sync.set({options: options});
+    await browser.storage.sync.set({ options: options });
     return;
 }
 
 function loadOptionsToDom(options) {
+    downloadPathTx.value = options.downloadPath;
     enableLoggingCB.checked = options.enableLogging;
     useCustomServerCB.checked = options.useCustomServer;
     customServerAddressTx.value = options.customServerAddress;
@@ -50,7 +59,7 @@ function loadOptionsToDom(options) {
 }
 
 function updateCustomServerAddressVisibility() {
-    customServerAddressTx.parentElement.style.display = useCustomServerCB.checked ? 'inherit': 'none';
+    customServerAddressTx.parentElement.style.display = useCustomServerCB.checked ? 'inherit' : 'none';
 }
 
 useCustomServerCB.addEventListener('change', updateCustomServerAddressVisibility)
@@ -61,5 +70,5 @@ optionsFrom.addEventListener('submit', e => {
 })
 
 resetDefaultBtn.addEventListener('click', _ => {
-    loadOptionsToDom(defaultOptions);
+    loadOptionsToDom(DEFAULT_OPTIONS);
 })
